requirements.txt
Django>=5.1  # Adjust the version based on your needs
djangorestframework>=3.14.0
drf-yasg>=1.21.4
python-dotenv>=1.0.0
psycopg2-binary>=2.9.8  # PostgreSQL database adapter for Python
